Fix warnings for XCode 9

+++ src/XCode/LocalNotificationLib/LocalNotificationTests/JKLegacyActionBuilderTest.m
@@ -44,7 +44,7 @@
 
 - (void)assertActions:(NSArray <UIUserNotificationAction *> *)results basedOnActions:(NSArray <JKLocalNotificationAction *> *)actions {
     XCTAssertEqual(results.count, actions.count);
-    for (int i = results.count - 1; i >= 0; i--) {
+    for (int i = (int)results.count - 1; i >= 0; i--) {
         [self assertAction:results[i] basedOnAction:actions[i]];
     }
 }
